Rename multiple refactor packages
Based on our conversations this changes the following:

- `bind` -> `bindshell` - Sometimes you need to bind things that aren't
  shells, this is done for clarity
- `download` -> `dropper` - Download is ambiguous and dropper is common
  jargon enough to help clarify
- `other` - Completely removed this and migrated it to be specific. The
  idea is that we would much rather have a bunch of smaller highly
  specific packages vs a giant catch all and reintroduce the same
  function pile problem this is trying to solve. More specific > less.
