+++ payload/bindshell/bindshell.go
@@ -1,18 +1,18 @@
 // Bind shell payloads & listeners.
 //
-// The download package contains all the code for generating bind shells that listen on a port for
+// The bindshell package contains all the code for generating bind shells that listen on a port for
 // interaction.
 //
 // As with all the payload sub-packages, each of these payload types can be used either in the raw string
 // format for manipulation or via the specific payload type provided by the project. This allows the
 // following usage patterns to have the same output:
 //
-//	bind.Netcat.Default(1337)
-//	bind.Netcat.Mknod(1337)
-//	fmt.Sprintf(bind.NetcatMknod, 1337)
-package bind
+//	bindshell.Netcat.Default(1337)
+//	bindshell.Netcat.Mknod(1337)
+//	fmt.Sprintf(bindshell.NetcatMknod, 1337)
+package bindshell
 
-type Bind interface {
+type BindShell interface {
 	Default
 }
