+++ payload/bindshell/bindshell_test.go
@@ -1,4 +1,4 @@
-package bind
+package bindshell
 
 import (
 	"strings"
