+++ payload/bindshell/telnet.go
@@ -1,4 +1,4 @@
-package bind
+package bindshell
 
 import (
 	"fmt"
