+++ payload/cron/cron.go
@@ -1,7 +1,7 @@
-// Other types of payloads that do not fit into the core categories.
+// cron based payloads.
 //
-// The payload package contains any specific payloads that do not fit into the other sub package types.
-package other
+// The payload package contains cron and cronjob based payloads.
+package cron
 
 import (
 	"fmt"
