+++ payload/cron/cron_test.go
@@ -1,4 +1,4 @@
-package other
+package cron
 
 import (
 	"testing"
