+++ payload/dropper/dropper.go
@@ -1,10 +1,10 @@
-// Download and execute payloads.
+// File dropper download and execute payloads.
 //
-// The download package contains all the code for download and execute payloads. Unlike the other payloads
+// The dropper package contains all the code for download and execute payloads. Unlike the other payloads
 // this package is necessarily OS dependent for both the download and execution portions.
-package download
+package dropper
 
-type Download interface{}
+type Dropper interface{}
 
 type UnixPayload struct{}
 type WindowsPayload struct{}
