+++ payload/dropper/dropper_test.go
@@ -1,4 +1,4 @@
-package download
+package dropper
 
 import (
 	"strings"
