+++ payload/dropper/unix.go
@@ -1,4 +1,4 @@
-package download
+package dropper
 
 import (
 	"fmt"
